requirements.txt
finnhub-python
pyTelegramBotAPI
requests
pytz
